Refactor: Separate model classes by functionality
