+++ TrackerGG/Models/Apex/apex_segment.py
@@ -2,7 +2,9 @@
 
 """
 Copyright (c) 2023 DevRuby
+
 MIT License
+
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -11,13 +13,20 @@ HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
 WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
+
 """
 
-from .abstract_http_client import AbstractHTTPClient
-from .httpclient import HTTPClientLibrary
-from .httpclient import RequestMethod
-from .httpclient import ResponseData
-from .httpclient import Route
-from .httpclient import get_http_client
+from typing import *
 
-__all__ = ["AbstractHTTPClient", "HTTPClientLibrary", "RequestMethod", "ResponseData", "Route", "get_http_client"]
+from .apex_stats import ApexStats
+
+__all__ = ["ApexSegment"]
+
+
+class ApexSegment:
+    def __init__(self, data: Dict[str, Union[str, dict]]):
+        self.type: str = data["type"]
+        self.attributes: dict = data["attributes"]
+        self.metadata: dict = data["metadata"]
+        self.expiry_date: str = data["expiryDate"]
+        self.stats: ApexStats = ApexStats(data["stats"])
