+++ TrackerGG/Models/CSGO/__init__.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+
+"""
+Copyright (c) 2023 DevRuby
+
+MIT License
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NON INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+OTHER DEALINGS IN THE SOFTWARE.
+
+"""
+
+from .csgo_map import CSGOMap
+from .csgo_map_segment import CSGOMapSegment
+from .csgo_map_stats import CSGOMapStats
+from .csgo_profile import CSGOProfile
+from .csgo_query_data import CSGOQueryData
+from .csgo_segment import CSGOSegment
+from .csgo_stats import CSGOStats
+from .csgo_weapon import CSGOWeapon
+from .csgo_weapon_segment import CSGOWeaponSegment
+from .csgo_weapon_stats import CSGOWeaponStats
+
+__all__ = [
+    "CSGOMap",
+    "CSGOMapSegment",
+    "CSGOMapStats",
+    "CSGOProfile",
+    "CSGOQueryData",
+    "CSGOSegment",
+    "CSGOStats",
+    "CSGOWeapon",
+    "CSGOWeaponSegment",
+    "CSGOWeaponStats",
+]
