+++ TrackerGG/Models/CSGO/csgo_weapon.py
@@ -1,0 +1,54 @@
+# -*- coding: utf-8 -*-
+
+"""
+Copyright (c) 2023 DevRuby
+
+MIT License
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NON INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+OTHER DEALINGS IN THE SOFTWARE.
+
+"""
+
+from enum import Enum
+
+__all__ = ["CSGOWeapon"]
+
+
+class CSGOWeapon(Enum):
+    AK47 = "ak47"
+    AUG = "aug"
+    AWP = "awp"
+    BIZON = "bizon"
+    DEAGLE = "deagle"
+    DUALS = "elite"
+    FAMAS = "famas"
+    FIVESEVEN = "fiveseven"
+    G3SG1 = "g3sg1"
+    GALIL = "galilar"
+    GLOCK = "glock"
+    P2000 = "hkp2000"
+    M249 = "m249"
+    M4A1 = "m4a1"
+    MAC10 = "mac10"
+    MAG7 = "mag7"
+    MP7 = "mp7"
+    MP9 = "mp9"
+    NEGEV = "negev"
+    NOVA = "nova"
+    P250 = "p250"
+    P90 = "p90"
+    SAWED_OFF = "sawedoff"
+    SCAR = "scar20"
+    SG553 = "sg556"
+    SSG08 = "ssg08"
+    ZEUS = "taser"
+    TEC9 = "tec9"
+    UMP45 = "ump45"
+    XM1014 = "xm1014"
