+++ TrackerGG/Models/General/platform.py
@@ -2,7 +2,9 @@
 
 """
 Copyright (c) 2023 DevRuby
+
 MIT License
+
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -11,13 +13,32 @@ HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
 WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
+
 """
 
-from .abstract_http_client import AbstractHTTPClient
-from .httpclient import HTTPClientLibrary
-from .httpclient import RequestMethod
-from .httpclient import ResponseData
-from .httpclient import Route
-from .httpclient import get_http_client
+from enum import Enum, auto
 
-__all__ = ["AbstractHTTPClient", "HTTPClientLibrary", "RequestMethod", "ResponseData", "Route", "get_http_client"]
+__all__ = ["Platform"]
+
+
+class Platform(Enum):
+    steam = auto()
+    origin = auto()
+    xbl = auto()
+    psn = auto()
+    uplay = auto()
+    battlenet = auto()
+
+    @staticmethod
+    def from_name(raw: str):
+        raw = raw.lower()
+
+        platform = {
+            "steam": Platform.steam,
+            "origin": Platform.origin,
+            "xbl": Platform.xbl,
+            "psn": Platform.psn,
+            "uplay": Platform.uplay,
+        }.get(raw)
+
+        return platform
