+++ TrackerGG/Models/General/social_account.py
@@ -16,7 +16,9 @@ OTHER DEALINGS IN THE SOFTWARE.
 
 """
 
-from typing import Dict, Union, Optional, Any, List
+from typing import *
+
+__all__ = ["SocialAccount"]
 
 
 class SocialAccount:
@@ -27,23 +29,3 @@ class SocialAccount:
         self.platform_user_identifier: Union[str, int] = data["platformUserIdentifier"]
         self.avatar_url: Optional[str] = data.get("avatarUrl")
         self.additional_parameters: Any = data["additionalParameters"]
-
-
-class UserInfo:
-    def __init__(self, data: Dict[str, Union[str, int, bool, list, None]]):
-        social_accounts = []
-        if data["socialAccounts"]:
-            for social_account in data["socialAccounts"]:
-                social_accounts.append(SocialAccount(social_account))
-
-        self.user_id: Optional[int] = data["userId"]
-        self.is_premium: bool = data["isPremium"]
-        self.is_verified: bool = data["isVerified"]
-        self.is_influencer: bool = data["isInfluencer"]
-        self.is_partner: bool = data["isPartner"]
-        self.country_code: Union[str, None] = data["countryCode"]
-        self.custom_avatar_url: Optional[str] = data["customAvatarUrl"]
-        self.custom_hero_url: Optional[str] = data["customHeroUrl"]
-        self.social_accounts: List[SocialAccount] = social_accounts
-        self.page_views: int = data["pageviews"]
-        self.custom_avatar_url: Any = data["isSuspicious"]
