--- TrackerGG/Models/csgo.py
@@ -1,159 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Copyright (c) 2023 DevRuby
-
-MIT License
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-NON INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-OTHER DEALINGS IN THE SOFTWARE.
-
-"""
-from enum import Enum
-from typing import Dict, Any, List, Union, Optional
-
-from .platform import PlatformInfo
-from .segment import Stat
-from .user import UserInfo
-
-
-class CSGOStats:
-    def __init__(self, data: Dict[str, dict]):
-        self.time_played: Stat = Stat(data["timePlayed"])
-        self.score: Stat = Stat(data["score"])
-        self.kills: Stat = Stat(data["kills"])
-        self.deaths: Stat = Stat(data["deaths"])
-        self.kd: Stat = Stat(data["kd"])
-        self.damage: Stat = Stat(data["damage"])
-        self.headshots: Stat = Stat(data["headshots"])
-        self.dominations: Stat = Stat(data["dominations"])
-        self.shots_fired: Stat = Stat(data["shotsFired"])
-        self.shots_hit: Stat = Stat(data["shotsHit"])
-        self.shots_accuracy: Stat = Stat(data["shotsAccuracy"])
-        self.snipers_killed: Stat = Stat(data["snipersKilled"])
-        self.domination_overkills: Stat = Stat(data["dominationOverkills"])
-        self.domination_revenges: Stat = Stat(data["dominationRevenges"])
-        self.bombs_planted: Stat = Stat(data["bombsPlanted"])
-        self.bombs_defused: Stat = Stat(data["bombsDefused"])
-        self.money_earned: Stat = Stat(data["moneyEarned"])
-        self.hostages_rescued: Stat = Stat(data["hostagesRescued"])
-        self.mvp: Stat = Stat(data["mvp"])
-        self.wins: Stat = Stat(data["wins"])
-        self.ties: Stat = Stat(data["ties"])
-        self.matches_played: Stat = Stat(data["matchesPlayed"])
-        self.losses: Stat = Stat(data["losses"])
-        self.rounds_played: Stat = Stat(data["roundsPlayed"])
-        self.rounds_won: Stat = Stat(data["roundsWon"])
-        self.wl_percentage: Stat = Stat(data["wlPercentage"])
-        self.headshot_pct: Stat = Stat(data["headshotPct"])
-
-
-class CSGOMapStats:
-    def __init__(self, data: Dict[str, dict]):
-        self.rounds: Stat = Stat(data["rounds"])
-        self.wins: Stat = Stat(data["wins"])
-
-
-class CSGOWeaponStats:
-    def __init__(self, data: Dict[str, dict]):
-        self.kills: Stat = Stat(data["kills"])
-        self.shots_fired: Stat = Stat(data["shotsFired"])
-        self.shots_hit: Stat = Stat(data["shotsHit"])
-        self.shots_accuracy: Stat = Stat(data["shotsAccuracy"])
-
-
-class CSGOMapSegment:
-    def __init__(self, data: Dict[str, Union[str, dict]]):
-        self.type: str = data["type"]
-        self.attributes: dict = data["attributes"]
-        self.metadata: dict = data["metadata"]
-        self.expiry_date: str = data["expiryDate"]
-        self.stats: CSGOMapStats = CSGOMapStats(data["stats"])
-
-
-class CSGOWeaponSegment:
-    def __init__(self, data: Dict[str, Union[str, dict]]):
-        self.type: str = data["type"]
-        self.attributes: dict = data["attributes"]
-        self.metadata = data["metadata"]
-        self.expiry_date: str = data["expiryDate"]
-        self.stats: CSGOWeaponStats = CSGOWeaponStats(data["stats"])
-
-
-class CSGOSegment:
-    def __init__(self, data: Dict[str, Union[str, dict]]):
-        self.type: str = data["type"]
-        self.attributes: dict = data["attributes"]
-        self.metadata: dict = data["metadata"]
-        self.expiry_date: str = data["expiryDate"]
-        self.stats: CSGOStats = CSGOStats(data["stats"])
-
-
-class CSGOProfile:
-    def __init__(self, data: Dict[str, Any]):
-        segments = []
-        for seg in data["segments"]:
-            segments.append(CSGOSegment(seg))
-
-        self.platform_info: PlatformInfo = PlatformInfo(data["platformInfo"])
-        self.user_info: UserInfo = UserInfo(data["userInfo"])
-        self.segments: List[CSGOSegment] = segments
-        self.expiry_date: str = data["expiryDate"]
-
-
-class CSGOMap:
-    def __init__(self, data: Dict[str, Union[str, dict]]):
-        self.attributes: dict = data["attributes"]
-        self.metadata: dict = data["metadata"]
-        self.expiry_date: str = data["expiryDate"]
-        self.stats: CSGOMapStats = CSGOMapStats(data["stats"])
-
-
-class CSGOQueryData:
-    def __init__(self, data: Dict[str, Optional[Union[str, int]]]):
-        self.platform_id = int(data["platformId"])
-        self.platform_user_id: Union[str, int] = data["platformUserId"]
-        self.platform_user_handle: str = data["platformUserHandle"]
-        self.platform_user_identifier: Union[str, int] = data["platformUserIdentifier"]
-        self.avatar_url: str = data["avatarUrl"]
-        self.additional_parameters: Any = data["additionalParameters"]
-        self.status: Any = data["status"]
-
-
-class CSGOWeapon(Enum):
-    AK47 = "ak47"
-    AUG = "aug"
-    AWP = "awp"
-    BIZON = "bizon"
-    DEAGLE = "deagle"
-    DUALS = "elite"
-    FAMAS = "famas"
-    FIVESEVEN = "fiveseven"
-    G3SG1 = "g3sg1"
-    GALIL = "galilar"
-    GLOCK = "glock"
-    P2000 = "hkp2000"
-    M249 = "m249"
-    M4A1 = "m4a1"
-    MAC10 = "mac10"
-    MAG7 = "mag7"
-    MP7 = "mp7"
-    MP9 = "mp9"
-    NEGEV = "negev"
-    NOVA = "nova"
-    P250 = "p250"
-    P90 = "p90"
-    SAWED_OFF = "sawedoff"
-    SCAR = "scar20"
-    SG553 = "sg556"
-    SSG08 = "ssg08"
-    ZEUS = "taser"
-    TEC9 = "tec9"
-    UMP45 = "ump45"
-    XM1014 = "xm1014"
